requirements.txt
Flask==2.3.3
pymongo==4.5.0
PyJWT==2.8.0
Werkzeug==2.3.7
python-dotenv==1.0.0
requests==2.31.0
flask-cors==4.0.0